# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presensi API",
        "description": "API documentation for the Presensi application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "/docs"
        }
    ],
    "paths": {
        "/yf/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "Login user yang sudah terdaftar",
                "operationId": "0a3b8667054db84c47c94f22db0965d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil login"
                    },
                    "401": {
                        "description": "Login gagal"
                    }
                }
            }
        },
        "/yf/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "description": "Logout user yang sedang login",
                "operationId": "b14648c1fbd6b339ce814c09df43d306",
                "responses": {
                    "200": {
                        "description": "Berhasil logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logout successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/yf/cuti/data-cuti": {
            "get": {
                "tags": [
                    "Cuti"
                ],
                "summary": "Memuat data cuti",
                "description": "Mendapatkan data cuti yang sedang login berdasarkan token autentikasi",
                "operationId": "7ba452588da43a1717d623d89522dda4",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data cuti",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "employee_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "employee_email": {
                                                    "type": "string",
                                                    "example": "johndoe@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data Cuti Tidak Ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Employee data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/yf/cuti/tambah-data-cuti": {
            "post": {
                "tags": [
                    "Cuti"
                ],
                "summary": "Memuat data cuti",
                "description": "Menambahkan data cuti yang sedang login berdasarkan token autentikasi",
                "operationId": "f1b76b39a0f793c51e57bf86c50f41c0",
                "responses": {
                    "200": {
                        "description": "Berhasil menyimpan data cuti",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "employee_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "employee_email": {
                                                    "type": "string",
                                                    "example": "johndoe@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}